+++ CHANGELOG.md
@@ -4,9 +4,9 @@
 
 ### Fixed
 
+- [CL-915] Show "Messaging" menu item only if any of 3 messaging features enabled
 - [CL-940] Graph excel export was not always giving the same numbers as the graph itself if a time period was selected. This is fixed now.
 - [CL-888] Fixed issue with folders page layout. Project cards are now displayed in two columns (instead of one) on large screens.
--
 
 ## 2022-06-13